requirements.txt
pip~=20.2.4
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
python-dotenv~=0.15.0
MarkupSafe~=1.1.1
certifi~=2020.6.20
chardet~=3.0.4
urllib3~=1.25.11
idna~=2.10
setuptools~=50.3.2
requests~=2.24.0
Flask~=1.1.2
gunicorn~=20.0.4